Cleanup & refactoring folder structure
