+++ spec/SelectableSystem_spec.lua
@@ -1,6 +1,6 @@
 inspect = require 'lib/inspect'
 SelectableSystem = require 'src/system/SelectableSystem' 
-Registry = require 'src/structure/Registry'
+Registry = require 'src/Registry'
 
 describe('SelectableSystem', function ()
 	it('should initialize correctly', function ()
