--- src/state/GameState.lua
@@ -1,7 +1,0 @@
---GameState.lua
-local class = require 'lib/30log'
-local GameState = class('GameState', {
-
-})
-
-return GameState
